requirements.txt
numpy 
scikit-learn 
wget 
matplotlib 
ipywidgets 
ipykernel
tensorflow
pillow
